[step3-pr] refactor: step3 패키지 삭제, 다른 패키지랑 겹치는 패스가 있어서 이전 구현체들 위치 변경(step1)
- pr : https://github.com/next-step/kotlin-racingcar/pull/482#discussion_r871377055
