+++ .gitignore
@@ -3,7 +3,7 @@ HELP.md
 build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**
-!**/src/test/**
+!**/src/step1.test/**
 
 ### STS ###
 .apt_generated
