--- src/main/kotlin/Person.kt
@@ -1,1 +1,0 @@
-data class Person(val name: String, val age: Int, val nickname: String? = name)
