+++ src/main/kotlin/racingcar/InputParser.kt
@@ -1,0 +1,8 @@
+package racingcar
+
+object InputParser {
+
+    fun toInt(input: String?): Int {
+        return input?.toIntOrNull() ?: throw IllegalArgumentException("잘못된 입력 입니다 `$input`")
+    }
+}
