+++ src/main/kotlin/racingcar/RandomNumberGenerator.kt
@@ -1,0 +1,9 @@
+package racingcar
+
+object RandomNumberGenerator {
+    private val NUMBER_RANGE = 1..9
+
+    fun generate(): Int {
+        return NUMBER_RANGE.random()
+    }
+}
