+++ src/main/kotlin/racingcar/car/Driver.kt
@@ -1,0 +1,19 @@
+package racingcar.car
+
+import racingcar.car.strategy.MoveStrategy
+
+class Driver(
+    private val car: Car,
+    private val moveStrategy: MoveStrategy
+) {
+
+    fun drive() {
+        if (moveStrategy.isMovable()) {
+            car.move()
+        }
+    }
+
+    fun getMoveResult(): Int {
+        return car.currentDrivingRecord
+    }
+}
