+++ src/main/kotlin/racingcar/car/Position.kt
@@ -1,0 +1,6 @@
+package racingcar.car
+
+enum class Position(desc: String) {
+    MOVE("이동"),
+    STOP("멈춤")
+}
