+++ src/main/kotlin/racingcar/car/exception/GameOverException.kt
@@ -1,0 +1,3 @@
+package racingcar.car.exception
+
+class GameOverException : RuntimeException("게임을 더이상 할 수 없습니다")
