+++ src/main/kotlin/racingcar/car/strategy/MoveStrategy.kt
@@ -1,0 +1,5 @@
+package racingcar.car.strategy
+
+interface MoveStrategy {
+    fun isMovable(): Boolean
+}
