+++ src/main/kotlin/racingcar/car/strategy/RandomNumberStrategy.kt
@@ -1,0 +1,13 @@
+package racingcar.car.strategy
+
+import racingcar.RandomNumberGenerator
+
+class RandomNumberStrategy : MoveStrategy {
+    override fun isMovable(): Boolean {
+        return RandomNumberGenerator.generate() > DRIVE_DEFINE_NUMBER
+    }
+
+    companion object {
+        private const val DRIVE_DEFINE_NUMBER = 4
+    }
+}
