+++ src/main/kotlin/step1/Person.kt
@@ -1,0 +1,3 @@
+package step1
+
+data class Person(val name: String, val age: Int, val nickname: String? = name)
