--- src/main/kotlin/step3/racingcar/InputParser.kt
@@ -1,8 +1,0 @@
-package step3.racingcar
-
-object InputParser {
-
-    fun toInt(input: String?): Int {
-        return input?.toIntOrNull() ?: throw IllegalArgumentException("잘못된 입력 입니다 `$input`")
-    }
-}
