--- src/main/kotlin/step3/racingcar/RandomNumberGenerator.kt
@@ -1,9 +1,0 @@
-package step3.racingcar
-
-object RandomNumberGenerator {
-    private val NUMBER_RANGE = 1..9
-
-    fun generate(): Int {
-        return NUMBER_RANGE.random()
-    }
-}
