--- src/main/kotlin/step3/racingcar/car/Car.kt
@@ -1,10 +1,0 @@
-package step3.racingcar.car
-
-class Car {
-    private val _drivingRecord: MutableList<Position> = mutableListOf()
-    val currentDrivingRecord get() = _drivingRecord.count { it == Position.MOVE }
-
-    fun move() {
-        _drivingRecord.add(Position.MOVE)
-    }
-}
