--- src/main/kotlin/step3/racingcar/car/Driver.kt
@@ -1,19 +1,0 @@
-package step3.racingcar.car
-
-import step3.racingcar.car.strategy.MoveStrategy
-
-class Driver(
-    private val car: Car,
-    private val moveStrategy: MoveStrategy
-) {
-
-    fun drive() {
-        if (moveStrategy.isMovable()) {
-            car.move()
-        }
-    }
-
-    fun getMoveResult(): Int {
-        return car.currentDrivingRecord
-    }
-}
