--- src/main/kotlin/step3/racingcar/car/Position.kt
@@ -1,6 +1,0 @@
-package step3.racingcar.car
-
-enum class Position(desc: String) {
-    MOVE("이동"),
-    STOP("멈춤")
-}
