--- src/main/kotlin/step3/racingcar/car/exception/GameOverException.kt
@@ -1,3 +1,0 @@
-package step3.racingcar.car.exception
-
-class GameOverException : RuntimeException("게임을 더이상 할 수 없습니다")
