--- src/main/kotlin/step3/racingcar/car/strategy/MoveStrategy.kt
@@ -1,5 +1,0 @@
-package step3.racingcar.car.strategy
-
-interface MoveStrategy {
-    fun isMovable(): Boolean
-}
