--- src/main/kotlin/step3/racingcar/view/ResultView.kt
@@ -1,15 +1,0 @@
-package step3.racingcar.view
-
-object ResultView {
-    private const val DASH_VIEW = "-"
-
-    fun showRecordsAsDash(driveRecords: List<Int>) {
-        driveRecords.forEach { record ->
-            repeat(record) {
-                print(DASH_VIEW)
-            }
-            println()
-        }
-        println()
-    }
-}
