+++ src/test/kotlin/racingcar/car/AlwaysTrueMoveStrategy.kt
@@ -1,0 +1,9 @@
+package racingcar.car
+
+import racingcar.car.strategy.MoveStrategy
+
+class AlwaysTrueMoveStrategy : MoveStrategy {
+    override fun isMovable(): Boolean {
+        return true
+    }
+}
