--- src/test/kotlin/step3/racingcar/car/AlwaysTrueMoveStrategy.kt
@@ -1,9 +1,0 @@
-package step3.racingcar.car
-
-import step3.racingcar.car.strategy.MoveStrategy
-
-class AlwaysTrueMoveStrategy : MoveStrategy {
-    override fun isMovable(): Boolean {
-        return true
-    }
-}
