--- src/test/kotlin/step3/racingcar/view/ViewTest.kt
@@ -1,11 +1,0 @@
-package step3.racingcar.view
-
-import org.junit.jupiter.api.Test
-
-class ViewTest {
-
-    @Test
-    fun `ResultView 객체를 생성, 입력값 만큼 -를 print한다`() {
-        ResultView.showRecordsAsDash(listOf(5))
-    }
-}
