--- src/test/kotlin/test/StringKoTest.kt
@@ -1,12 +1,0 @@
-package test
-
-import io.kotest.core.spec.style.AnnotationSpec
-import io.kotest.matchers.shouldBe
-
-internal class StringKoTest : AnnotationSpec() {
-
-    @Test
-    fun isEmpty() {
-        "".isEmpty() shouldBe true
-    }
-}
